Ability to not intercept touch events
This PR introduces a new attribute, `interceptTouchEvents`, that can be used to prevent `TelescopeLayout` from consuming touch events.

The default value is `true`, which means that by default `TelescopeLayout` will consume touch events when two fingers are on the screen. So the default behavior hasn't changed.

However, when `interceptTouchEvents` is false, views under `TelescopeLayout` will still receive touch events when two fingers are on the screen. This means that for instance two finger zooming now works, even though Telescope is also going to start in parallel.