Bitmap Processor support is added.
What I did was to split capturing of the screenshot and saving it. `onCapture` with Bitmap parameter is called first in the Lens. Possible bitmap transformations can be made in that method. After the operation finishes, `onBitmapReady` method of `BitmapProcessorListener` will be called. The default implementation does nothing and just calls `onBitmapReady`.

This processing operation can take long. That's why I did it like this to inform TelescopeLayout to continue the operation when it finishes. While this operation continues, `capturing` is kept `true` and all the screen touch events are ignored.

I don't like introducing new classes into other people's libraries. Sorry about that.

This worked for me for my Google Maps screenshot problem. I opened the PR to discuss it. I also want to create Maps Activity in the sample to demonstrate it.

Thank you for the great library. Thank you for taking time to review this.