Example tab with Google Maps canvas screenshot is added.
Google works so hard for people to use `MapFragment`. Docs say that if we are using `MapView` we have to call 6 lifecycle callbacks manually. Which I am not doing here. I didn't take time to do it cleanly. For demo purposes I think it is fine. What do you think?

For maps to work, we need API key. So I included a `debug.keystore` into the source code with a API key taken from Google Developers Cloud Console. It is not the best practice to put this API key into source control but for demo purposes I think it is fine. You may want to change that.