Adhere to the API 34 media projection requirements (#107)
There is a strict ordering requirement starting with API 34. A media
projection service cannot be started until after the permission has
been granted. With the service starting being deferred, we can no
longer rely on it being started by the time we start the projection in
the first broadcast receiver and must wait for another broadcast.